Finish A-Packages
Put all the classes in diego folder
