+++ src/main/java/diego/AddDeadlineCommand.java
@@ -1,13 +1,16 @@
-public class AddTodoCommand implements Command {
+package diego;
+public class AddDeadlineCommand implements Command {
     private String description;
+    private String by;
 
-    public AddTodoCommand(String description) {
+    public AddDeadlineCommand(String description, String by) {
         this.description = description;
+        this.by = by;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        Task task = new Todo(description);
+        Task task = new Deadline(description, by);
         tasks.add(task);
         ui.showTaskAdded(task, tasks.size());
         storage.save(tasks);
