+++ src/main/java/diego/Command.java
@@ -1,3 +1,4 @@
+package diego;
 public interface Command {
     void execute(TaskList tasks, Ui ui, Storage storage);
 }
