+++ src/main/java/diego/Deadline.java
@@ -1,3 +1,4 @@
+package diego;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
