+++ src/main/java/diego/Diego.java
@@ -1,3 +1,4 @@
+package diego;
 import java.io.FileNotFoundException;
 import java.util.Scanner;
