+++ src/main/java/diego/DiegoException.java
@@ -1,3 +1,4 @@
+package diego;
 public class DiegoException extends Exception{
     public DiegoException(String message){
         super(message);
