+++ src/main/java/diego/Event.java
@@ -1,3 +1,4 @@
+package diego;
 public class Event extends Task {
     protected String start;
     protected String end;
