+++ src/main/java/diego/ExitCommand.java
@@ -1,3 +1,4 @@
+package diego;
 public class ExitCommand implements Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
