+++ src/main/java/diego/ListCommand.java
@@ -1,3 +1,4 @@
+package diego;
 public class ListCommand implements Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
