+++ src/main/java/diego/Storage.java
@@ -1,3 +1,4 @@
+package diego;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
