+++ src/main/java/diego/Task.java
@@ -1,3 +1,4 @@
+package diego;
 public abstract class Task {
     protected String description;
     protected boolean isDone;
